Fix for Bug 948047 and Issue 40.
For some reason the ConnectorManagerGetServlet.doGet() method was ignoring the given lang and passing null down to the processDoGet() abstract method.  This class is specialized by 3 other classes.  For 2 of those classes this didn't matter because they didn't return any locale specific resources.  However, the GetConnectorConfigToEdit servlet does need this information or else it uses the default locale.  Since the Connector Manager is running in a different environment than the GSA this leads to problems.

The GSA does send the Lang attribute with the request to getConnecotrConfigToEdit, so this fix allows that attribute to be passed down to the ConnectorType.getPopulatedConfigForm() SPI method.