Merged /trunk:r714 - Fix for Bug 950013 where Japanese characters did not display correctly (????) on adding a new connector.
Turns out the GetConfigForm Servlet did not set the encoding for the response
and it defaulted to ISO-8859-1.  This change sets the encoding to UTF-8 so the
Unicode is properly transmitted back to the GSA.

I have also made a similar change to the UpdateConnector Servlet, however, this
Servlet doesn't appear to be used by the GSA Admin.