Merged /trunk:r719 Make PrefsStore Configurable
Changed the PerfsStore constructor to allow the caller to select between using
the userNode or the systemNode as the root node.  Also added a flush() method to
support testing persisting the prefs.

No changes are needed to the Connector Manager configuration and it defaults to
the userNode as previously implemented.

CAUTION: If you want to use the systemNode make sure the user account running
Tomcat has write access to the systemNode area.  If this is not the case you
will see the following warnings in the log file:

  WARNING: Couldn't flush system prefs: java.util.prefs.BackingStoreException: 
    /etc/.java/.systemPrefs/com create failed.

In order to change the Connector Manager to use the systemNode root edit the
"ConnectorScheduleStore" and "ConnectorStateStore" beans in the
webapps/connector-manager/WEB-INF/applicationContext.xml file as follows:

	<bean id="ConnectorScheduleStore"
		class="com.google.enterprise.connector.persist.PrefsStore">
		<constructor-arg index="0" type="boolean" value="false"/>
	</bean>

	<bean id="ConnectorStateStore"
		class="com.google.enterprise.connector.persist.PrefsStore">
		<constructor-arg index="0" type="boolean" value="false"/>
	</bean>