Merged /trunk:r730 Refactor DocPusher to XmlUtils
Refactor some of the XML related utility methods out of DocPusher into
a new class so they can easily be used in the GSA Admin similar to the
way ServletUtils are used.  This is the first step toward adding XML
encoding to the GSA Admin when returning the Connector Config Form.