I've tested on linux with this patch and it seems to shutdown every time.
In the WorkQueue.shutdown() method, it interruts all the WorkThreads and tries
to shutdown the LifeThread but it never really waits for the LifeThread to be
done.  On top of this, it does not remove any of the interrupted threads from
the 'threads' HashSet.  Therefore, if the LifeThread happens to be in the wait()
state when shutdown is called, it will check the status of all the treads in the
'threads' HashSet.  If it gets a chance to run before the Tomcat server is
shutdown and the JVM exits, then it will restart all the interrupted WorkThreads
and the Tomcat server will not be able to exit.

This change basically kills the LifeThread and makes sure it is killed before
taking care of all the WorkThreads.