This set of changes addresses Connector Manager Issue 81 - Encrypt all properties with (case insensitive) 'password' in the key.
This makes no attempt to encrypt password properties in the spring XML files.
It also makes no attempt to make sense of the property key name, so 
  useTrespassWordList=false
will get encrypted simply because it contains the substring 'password'.

As a side effect, we now version the properties files with the version of 
the CM that wrote the file.  This makes it easier to determine which
properties to decrypt vs. not.


Change Log:
----------

M  connector-manager/source/java/com/google/enterprise/connector/instantiator/InstanceInfo.java
   - Calls EncryptedPropertyPlaceholderConfigurer to encrypt and decrypt properties.   

M  connector-manager/source/java/com/google/enterprise/connector/instantiator/EncryptedPropertyPlaceholderConfigurer.java
   - Moved the logic to encrypt and decrypt password properties here,
     rather than duplicate it here and in InstanceInfo.
   - Uses a version stamp added to the Propeties, that allows us to diffentiate
     older style password encryption and newer style password encryption.