This set of changes addresses Connector Manager Issue 105_part2 - Allow retrieval of CM and/or Connector config by the GSA
Admin servlet to retrieve the configuration of the Connector Manager
and all Connector instances.  This servlet returns a ZIP archive of all
the configuration files.  Access to this servlet is restricted to either
localhost or gsa.feed.host, based upon the HTTP RemoteAddress.

Usage:
To retrieve the zipped configuration archive:
  http://[cm_host_addr]/connector-manager/getConfig
or
  http://[cm_host_addr]/connector-manager/getConfig/configuration.zip

Redirects and curl:
When requesting the configuration files, this servlet returns a redirect
to the actual ZIP filename, configuration.zip.
This allows the browser, wget, or curl to pull the true filename off
the redirected URL so that it can name the file when storing locally.
If using curl to retrieve the files, please to use 'curl -L' to tell
curl to follow the redirect.  Unfortunately, when using 'curl -O' to
save the file locally, curl uses the pre-redirected name, rather than
the post-redirected name when naming the local file.  This forces you
to use 'curl -L -o output_filename'.

Wget handles redirects appropriately without intervention, and names
the saved file as expected.

Change Log:
----------
A  projects/connector-manager/source/java/com/google/enterprise/connector/servlet/GetConfig.java
   - New admin servlet aimed a support personel.  
     It writes to the output a ZIP archive containing the connector manager
     and connector instance configuration files.

M  projects/connector-manager/source/webdocs/web.xml
   - Add getConfig command to GetConfig servlet.