Add Event publishing method to Context and test using ApplicationListener bean registered through the ApplicationContext.
First step toward adding additional services to the Connector Manager.  Rather
than tightly integrating them as a package of SPI Interfaces and Mock and/or
Simple implementations I'm adding the ability for components that have access to
the Context to send events that will be multi-cast to all listeners.

This change adds a new method to the Context to support this and adds a test
using this method with an ApplicationListener instantiated from a bean through
the normal use of the Spring ApplicationContext.

NOTE: Concern here is the exposure of the ApplicationEvent class in the method
signature.  To date, we only expose Spring classes to the testing code.  If this
is a concern I can create a ConnectorEvent class that wraps and abstracts the
Spring ApplicationEvent class.

To create listeners, simply add a bean to your applicationContext.xml file that
defines a class that implements the ApplicationListener interface.  Once the
context is refreshed all listeners will be registered.  Any events that are sent
to the Context will be multi-cast to all listeners.  Right now it is not possible
to filter events so all listeners will have to do that within the
onApplicationEvent() processing.  Note, it is not recommended that you add or
remove listeners during the life of the ApplicationContext.  It's best to
establish your listeners during initialization and leave them.