Summary:
Add a new ContextService interface to the manager package, a reserved name
TreeMap Spring Bean to be used to order ContextServices, added the usage of the
new interface and bean to the Context class, and added associated tests.


Objective:

To support adding services to the Connector Manager for the Hosted Enterprise
Search (HES) project, two main goals must be achieved:

1. Provide a way to easily add new services to the Connector Manager.  This has
   to be done in a manner that enables services that may depend on other
   services to declare the order in which they are started.
2. Provide a way to easily develop handlers to interact with these new services.
   Optionally, it would be nice to also allow Connectors to interact with the
   services outside of the SPI. 


Design:

Related to this change, it's important to understand a couple of aspects.  In
the current design, the Context class is the container for static services used
to establish the Spring ApplicationContext.  This consists of configuration,
registration, instantiation, and high-level startup and shutdown.  It currently
wraps the ApplicationContext and provides general methods to get Spring beans
from the associated BeanFactory and a specific method to get the Manager.  It
also currently conditionally starts/stops one service which is the Scheduler
service.

The new ContextService interface can be used to register services with the
Context.  They will be started and stopped along with the Context in a declared
order.

At a high-level, a developer can add a service like this:

1. Implement the ContextService interface.
2. Add a <bean> to the applicationContext.xml file to declare the
   ContextService and any properties needed.
3. If the ContextService has a load dependency it needs to be added to the
   OrderedServices making sure the key to the map entry will order the service
   properly.

The Context would load and start any ContextServices it finds in the
ApplicationContext.


Follow On Work:

Next change will try to convert the current Scheduler into a ContextService.