Added code to obfuscate sensitive data being sent during update operations.
NOTES TO DEVELOPERS:

The formSnippet returned from the ConnectorType methods MUST BE WELL FORMED
HTML.  In the past, the formSnippet was simply sent to the GSA for display in
the browser and the browser parser is more forgiving.  With the addition of this
new filter on the formSnippet, the formSnippet is now passed through an XML
parser and malformed HTML will not be passed back to the caller.  This means a
Connector can never be updated if the ConnectorType.getPopulatedConfigForm()
implementation is returning a malformed HTML formSnippet in the
ConfigureResponse.  An XML parsing error is returned instead.

I have seen several malformed formSnippets in the current connectors so
developers should review their code and test the formSnippets in an XML parser
or with this patch.


TODO:

- Look into using a more forgiving HTML/DOM parser rather than the XML parser
  that requires the form snippet to be well formed.  I looked into JTidy,
  however, it does not support HTML fragments.  I also looked at nekoHTML which
  does support fragments, however, it extends Xerces and will cause ClassLoader
  issues.


TESTS PASSED:

- All Unit Tests
- Tested with GSA.  Created connector and updated several times.  Observed tcp
  traffic and confirmed sensitive data obfuscated.


CHANGE LOG:

  M java/com/google/enterprise/connector/instantiator/EncryptedPropertyPlaceholderConfigurer.java
    - Added use of new SecurityUtils.

  M java/com/google/enterprise/connector/servlet/ServletUtil.java
    - Added utility methods to filter sensitive data from a form snippet and
      replace sensitive data in a config map.

  M java/com/google/enterprise/connector/servlet/ConnectorManagerGetServlet.java
    - Now process the form snippet first to remove sensitive data.
    - Changed flow to check for possible XML error before using given status.
      If the config form was invalid HTML it will no longer be returned.

  M java/com/google/enterprise/connector/servlet/SetConnectorConfigHandler.java
    - Added call to replace sensitive data on an update.

  M javatests/com/google/enterprise/connector/servlet/ServletUtilTest.java
    - Added tests for new utility methods.

  M javatests/com/google/enterprise/connector/servlet/GetConfigFormTest.java
    - Fixed formSnippet to be valid HTML.

  A java/com/google/enterprise/connector/common/SecurityUtils.java
  A javatests/com/google/enterprise/connector/common/SecurityUtilsTest.java
    - New class to hold some of the utility methods used by other packages.