This set of changes addresses Connector Manager Issue 133 - InstanceMap.createNewConnector() should delete partial connector directory if instantiation fails.
InstanceMap.createNewConnector() now deletes the directory it
created if the instantiation fails.  This required refactoring
InstanceMap.removeConnector() to extract the directory removal code.

As it stands, if the customer modified the connectorInstance.xml
file during the failed configuration, the file will not be removed.
But that is probably a good thing.

We also detect that no properties file exists and stop instantiation there,
throwing an InstanceInfoException("Connector configuration not found...").
This is much more informative that the Spring PropertyPlacholderException
that was thrown previously.

Finally, if on startup, a connector does not have a schedule and
none was found in a legacy store, I create a disabled schedule and
log a warning.  This is better than spamming the log file with
"Connector Schedule not found" messages once a second.


Change Log:
----------
M  projects/connector-manager/source/java/com/google/enterprise/connector/instantiator/InstanceInfo.java
   - Throw exception if connector configuration not found.
   - Create a disabled schedule for connectors that are missing one.

M  projects/connector-manager/source/java/com/google/enterprise/connector/instantiator/InstanceMap.java
   - createNewConnector() deletes the directory it created
     if the instantiation fails, cleaning up detritus.