Description:
Fix for Bug 1721179.  Discovered that when the Test Connectors are used to send 
binary content over to the GSA the encoding gets mangled and the GSA cannot
decode the content for indexing.  Test Connectors are configured as a
JcrConnector.  After detailed investigation, discovered that the
JcrConnector.getValues() call was just pulling all values as a String.  This was
causing the InputStream to be read into a String only to be read back into a
Stream later.  Reading an InputStream into a String causes the original
character encoding to be mangled.

This fix removes the conversion from Stream to String when getting the values
for the case where the MockRepositoryProperty is not a multivalue property.

Tests Passed:

  - All Unit Tests - under JDK 1.5.
  - Tested WAR with GSA and used modified MockRepositoryEventLog4.txt to send
    both a .doc and .pdf to the GSA using the Test Connector.  Both are indexed
    and decoded fine.

ChangeLog:

M source/javatests/com/google/enterprise/connector/common/Base64FilterInputStreamTest.java:
  - Added a test for encoding the test.doc file using a JcrConnector.

M source/javatests/com/google/enterprise/connector/mock/jcr/MockJcrValue.java:
  - Added support for converting internal STREAM type to JCR BINARY type.

M source/javatests/com/google/enterprise/connector/mock/jcr/MockJcrProperty.java
  - Changed getValues() to return simple wrapped property in the case where
    multivalue is not an issue.  The problem with converting everything to a
    String is Stream encoding gets mangled.