Initial implementation of a thread pool built using java.util.conncurent components for use as a replacement for WorkQueue. This change includes modifications to TraverSalScheduler to use the new ThreadPool.
In addition the change slightly reorgainzies the logic in TraversalScheduler to
acquire Instantiator resources in the main scheduler thread rather than in the 
thread running the batch. My hope in this change is to help clarify the
resources the running batch requires and to enable testing of batch processing
without a dependency on an Instantiator.